requirements.txt
requests
aiohttp
pycryptodome
PyJWT
ujson
protobuf
protobuf-decoder
psutil
pytz
cfonts